`summary()` method for `fastglm` fitted objects

`summary()` method for `fastglm` fitted objects

# S3 method for class 'fastglm'
summary(
  object,
  dispersion = NULL,
  correlation = FALSE,
  symbolic.cor = FALSE,
  ...
)

Arguments

object

`fastglm` fitted object

dispersion

the dispersion parameter for the family used. Either a single numerical value or `NULL` (the default), when it is inferred from `object`.

correlation

logical; if `TRUE`, the correlation matrix of the estimated parameters is returned.

symbolic.cor

logical; if `TRUE`, print the correlations in a symbolic form (see `symnum`) rather than as numbers.

...

not used

Value

A `summary.fastglm` object

See also

Examples

x <- matrix(rnorm(10000 * 10), ncol = 10)
y <- 1 * (0.25 * x[,1] - 0.25 * x[,3] > rnorm(10000))

fit <- fastglm(x, y, family = binomial())

summary(fit)
#> 
#> Call:
#> fastglm.default(x = x, y = y, family = binomial())
#> 
#> Coefficients:
#>      Estimate Std. Error z value Pr(>|z|)    
#> X1   0.430087   0.021778  19.748   <2e-16 ***
#> X2  -0.003827   0.021031  -0.182    0.856    
#> X3  -0.453793   0.021656 -20.955   <2e-16 ***
#> X4  -0.019732   0.020998  -0.940    0.347    
#> X5  -0.008936   0.020556  -0.435    0.664    
#> X6   0.009996   0.020965   0.477    0.634    
#> X7  -0.008167   0.021143  -0.386    0.699    
#> X8   0.029878   0.020820   1.435    0.151    
#> X9  -0.024408   0.020823  -1.172    0.241    
#> X10 -0.013253   0.020674  -0.641    0.522    
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> (Dispersion parameter for binomial family taken to be 1)
#> 
#>     Null deviance: 13863  on 10000  degrees of freedom
#> Residual deviance: 12981  on  9990  degrees of freedom
#> AIC: 13001
#> 
#> Number of Fisher Scoring iterations: 4
#>